Add/update integration tests
