+++ sdk/python/tests/integration/test_langchain.py
@@ -3,35 +3,17 @@
 Integration tests for LangChain with TrainLoop LLM Logging.
 """
 
-import functools
 import os
 import pytest
-
-# Import TrainLoop SDK
-import trainloop_llm_logging as tl
+from pydantic import SecretStr
 
 # Test harness
-from .test_real_api_calls import (
+from .harness import (
     IntegrationTestHarness,
     require_openai_key,
     require_anthropic_key,
+    require_library,
 )
-
-
-def require_library(library_name: str):
-    """Decorator to skip tests if a library is not available."""
-
-    def decorator(func):
-        @functools.wraps(func)
-        def wrapper(*args, **kwargs):
-            try:
-                __import__(library_name)
-                return func(*args, **kwargs)
-            except ImportError:
-                pytest.skip(f"{library_name} library not installed. Install with: pip install {library_name}")
-        return wrapper
-
-    return decorator
 
 
 class TestLangChainIntegration:
@@ -42,26 +24,24 @@ class TestLangChainIntegration:
     def test_langchain_openai(self):
         """Test LangChain with OpenAI."""
         with IntegrationTestHarness("langchain_openai") as harness:
-            try:
-                from langchain_openai import ChatOpenAI
-            except ImportError:
-                pytest.skip("langchain_openai library not available")
+            from langchain_openai import ChatOpenAI
+
+            # Get API key
+            openai_api_key = os.getenv("OPENAI_API_KEY")
+            if not openai_api_key:
+                pytest.skip("OPENAI_API_KEY not set")
             # Create LangChain LLM
             llm = ChatOpenAI(
                 model="gpt-4o-mini",
-                openai_api_key=os.getenv("OPENAI_API_KEY"),
-                max_tokens=10,
+                api_key=SecretStr(openai_api_key),
+                max_completion_tokens=10,
             )
-
             # Make a call
             response = llm.invoke("Say hello in 3 words")
-
             assert response.content
-
             # Wait for JSONL entries
             entries = harness.wait_for_entries(expected_count=1)
             assert len(entries) >= 1, f"Expected at least 1 entry, got {len(entries)}"
-
             # Validate entry
             entry = entries[0]
             assert harness.validate_entry(entry, expected_model="gpt-4o-mini")
@@ -72,26 +52,25 @@ class TestLangChainIntegration:
     def test_langchain_anthropic(self):
         """Test LangChain with Anthropic."""
         with IntegrationTestHarness("langchain_anthropic") as harness:
-            try:
-                from langchain_anthropic import ChatAnthropic
-            except ImportError:
-                pytest.skip("langchain_anthropic library not available")
+            from langchain_anthropic import ChatAnthropic
+
+            # Get API key
+            anthropic_api_key = os.getenv("ANTHROPIC_API_KEY")
+            if not anthropic_api_key:
+                pytest.skip("ANTHROPIC_API_KEY not set")
             # Create LangChain LLM
             llm = ChatAnthropic(
-                model="claude-3-haiku-20240307",
-                anthropic_api_key=os.getenv("ANTHROPIC_API_KEY"),
-                max_tokens=10,
+                model_name="claude-3-haiku-20240307",
+                api_key=SecretStr(anthropic_api_key),
+                timeout=10,
+                stop=None,
             )
-
             # Make a call
             response = llm.invoke("Say hello in 3 words")
-
             assert response.content
-
             # Wait for JSONL entries
             entries = harness.wait_for_entries(expected_count=1)
             assert len(entries) >= 1, f"Expected at least 1 entry, got {len(entries)}"
-
             # Validate entry
             entry = entries[0]
             assert harness.validate_entry(
